+++ TelegramSearchBot.LLM/Service/AI/LLM/OpenAIService.cs
@@ -832,7 +832,7 @@ namespace TelegramSearchBot.Service.AI.LLM {
                                 var argsDict = JsonConvert.DeserializeObject<Dictionary<string, string>>(argsJson)
                                     ?? new Dictionary<string, string>();
 
-                                var toolContext = new ToolContext { ChatId = ChatId, UserId = message.FromUserId, MessageId = (int)message.MessageId };
+                                var toolContext = new ToolContext { ChatId = ChatId, UserId = message.FromUserId, MessageId = message.MessageId };
                                 object toolResultObject = await McpToolHelper.ExecuteRegisteredToolAsync(toolName, argsDict, toolContext);
                                 toolResultString = McpToolHelper.ConvertToolResultToString(toolResultObject);
                                 _logger.LogInformation("{ServiceName}: Tool {ToolName} executed. Result length: {Length}", ServiceName, toolName, toolResultString.Length);
@@ -860,7 +860,7 @@ namespace TelegramSearchBot.Service.AI.LLM {
                     executionContext.IterationLimitReached = true;
                     executionContext.SnapshotData = new LlmContinuationSnapshot {
                         ChatId = ChatId,
-                        OriginalMessageId = (int)message.MessageId,
+                        OriginalMessageId = message.MessageId,
                         UserId = message.FromUserId,
                         ModelName = modelName,
                         Provider = "OpenAI",
@@ -940,7 +940,7 @@ namespace TelegramSearchBot.Service.AI.LLM {
                         string toolResultString;
                         bool isError = false;
                         try {
-                            var toolContext = new ToolContext { ChatId = ChatId, UserId = message.FromUserId, MessageId = (int)message.MessageId };
+                            var toolContext = new ToolContext { ChatId = ChatId, UserId = message.FromUserId, MessageId = message.MessageId };
                             object toolResultObject = await McpToolHelper.ExecuteRegisteredToolAsync(parsedToolName, toolArguments, toolContext);
                             toolResultString = McpToolHelper.ConvertToolResultToString(toolResultObject);
                             _logger.LogInformation("{ServiceName}: Tool {ToolName} executed. Result: {Result}", ServiceName, parsedToolName, toolResultString);
@@ -967,7 +967,7 @@ namespace TelegramSearchBot.Service.AI.LLM {
                     executionContext.IterationLimitReached = true;
                     executionContext.SnapshotData = new LlmContinuationSnapshot {
                         ChatId = ChatId,
-                        OriginalMessageId = (int)message.MessageId,
+                        OriginalMessageId = message.MessageId,
                         UserId = message.FromUserId,
                         ModelName = modelName,
                         Provider = "OpenAI",
