+++ TelegramSearchBot/Service/Tools/SendPhotoToolService.cs
@@ -31,8 +31,8 @@ namespace TelegramSearchBot.Service.Tools {
         /// Resolves the effective reply-to message ID. Uses the explicitly provided value if set,
         /// otherwise falls back to the ToolContext.MessageId (the original user message).
         /// </summary>
-        private static ReplyParameters GetReplyParameters(int? explicitReplyToMessageId, ToolContext toolContext) {
-            var messageId = explicitReplyToMessageId ?? (toolContext.MessageId != 0 ? toolContext.MessageId : (int?)null);
+        private static ReplyParameters GetReplyParameters(long? explicitReplyToMessageId, ToolContext toolContext) {
+            var messageId = explicitReplyToMessageId ?? (toolContext.MessageId != 0 ? toolContext.MessageId : (long?)null);
             return messageId.HasValue ? new ReplyParameters { MessageId = messageId.Value } : null;
         }
 
@@ -41,7 +41,7 @@ namespace TelegramSearchBot.Service.Tools {
             [BuiltInParameter("The base64 encoded image data (without the data URI prefix).")] string base64Data,
             ToolContext toolContext,
             [BuiltInParameter("Optional caption for the photo (max 1024 characters).", IsRequired = false)] string caption = null,
-            [BuiltInParameter("Optional message ID to reply to.", IsRequired = false)] int? replyToMessageId = null) {
+            [BuiltInParameter("Optional message ID to reply to.", IsRequired = false)] long? replyToMessageId = null) {
             try {
                 byte[] imageBytes;
                 try {
@@ -94,7 +94,7 @@ namespace TelegramSearchBot.Service.Tools {
             [BuiltInParameter("The file path to the image on the server.")] string filePath,
             ToolContext toolContext,
             [BuiltInParameter("Optional caption for the photo (max 1024 characters).", IsRequired = false)] string caption = null,
-            [BuiltInParameter("Optional message ID to reply to.", IsRequired = false)] int? replyToMessageId = null) {
+            [BuiltInParameter("Optional message ID to reply to.", IsRequired = false)] long? replyToMessageId = null) {
             try {
                 if (string.IsNullOrWhiteSpace(filePath)) {
                     return new SendPhotoResult {
